backend/requirements.txt
PyJWT==2.12.1
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.9.2
requests==2.32.3

# Audio processing
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4
scipy==1.13.1

# ML models (torch/torchaudio installed separately in Dockerfile as CPU-only)
openai-whisper==20250625
whisper-timestamped==1.15.9
silero-vad==5.1.2

# NLP
spacy==3.7.6
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
